rename file to locator
