+++ packages/sdk/src/crud/workflows.ts
@@ -1,5 +1,5 @@
 import { MCPClient } from "../fetcher.ts";
-import { Workspace } from "../workspace.ts";
+import { Workspace } from "../locator.ts";
 
 export interface WorkflowStartParams {
   workflowName: string;
